requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
black==21.9b0
chardet==4.0.0
click==8.0.1
discord-py-interactions==3.0.2
discord.py==1.7.3
idna==3.2
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
python-dotenv==0.19.0
regex==2021.9.24
tomli==1.2.1
typing-extensions==3.10.0.2
yarl==1.6.3
